# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Campaign API",
        "description": "API documentation for Campaign management system",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/campaign-sources": {
            "get": {
                "tags": [
                    "Campaign Sources"
                ],
                "summary": "Get all campaign-source associations",
                "operationId": "0b209dd873c578140a712c4e92069a88",
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            },
            "post": {
                "tags": [
                    "Campaign Sources"
                ],
                "summary": "Create campaign-source association",
                "operationId": "ad67e03ca3682a158d4f7b0d267a7d6c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "campaign_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "source_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/campaign-sources/{id}": {
            "get": {
                "tags": [
                    "Campaign Sources"
                ],
                "summary": "Get campaign-source association by ID",
                "operationId": "79b9d8a809de2dc509b914ab015194da",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Campaign Sources"
                ],
                "summary": "Delete campaign-source association",
                "operationId": "b5212c8bbc094a34d8956c8d2446d545",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
        "/campaigns/{campaignId}/sources": {
            "get": {
                "tags": [
                    "Campaign Sources"
                ],
                "summary": "Get sources for a campaign",
                "operationId": "d11898a72a2d23e12fe755cdde04704c",
                "parameters": [
                    {
                        "name": "campaignId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/news-sources/{sourceId}/campaigns": {
            "get": {
                "tags": [
                    "Campaign Sources"
                ],
                "summary": "Get campaigns for a source",
                "operationId": "f32ae61cfd075ceed382588d722b231c",
                "parameters": [
                    {
                        "name": "sourceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/campaigns/{campaignId}/sources/{sourceId}": {
            "post": {
                "tags": [
                    "Campaign Sources"
                ],
                "summary": "Attach source to campaign",
                "operationId": "3ed1e084e21473fe64649e30b813c0ff",
                "parameters": [
                    {
                        "name": "campaignId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sourceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Campaign Sources"
                ],
                "summary": "Detach source from campaign",
                "operationId": "cc4760c02b9f46b0c7ce5cd16accade9",
                "parameters": [
                    {
                        "name": "campaignId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "sourceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        },
...
    },
...
}